@@ -2,7 +2,6 @@ apply plugin: 'com.android.application'
 
 android {
     compileSdkVersion androidCompileSdkVersion
-
     defaultConfig {
         applicationId "com.xiaofeng.androidlibs"
         minSdkVersion androidMinSdkVersion
@@ -18,25 +17,18 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
-    buildToolsVersion '27.0.3'
 }
 
 dependencies {
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
-    implementation 'com.android.support:support-vector-drawable:27.1.1'
-    implementation fileTree(include: ['*.jar'], dir: 'libs')
-    testImplementation 'junit:junit:4.12'
-    implementation(group: 'com.android.support', name: 'design', version: supportLibVersion)
-    implementation(group: 'com.android.support', name: 'appcompat-v7', version: supportLibVersion)
-    implementation(group: 'com.android.support', name: 'recyclerview-v7', version: supportLibVersion)
-    implementation(group: 'com.android.support', name: 'recyclerview-v7', version: supportLibVersion)
-
     implementation project(":flowlayoutmanager")
 
-    implementation 'us.feras.mdv:markdownview:1.1.0'
-    implementation 'com.android.support:support-v4:27.1.1'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'androidx.vectordrawable:vectordrawable:1.1.0'
+    implementation(group: 'com.google.android.material', name: 'material', version: supportLibVersion)
+    implementation(group: 'androidx.appcompat', name: 'appcompat', version: supportLibVersion)
+    implementation(group: 'androidx.recyclerview', name: 'recyclerview', version: supportLibVersion)
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 }
