@@ -4,8 +4,8 @@ apply plugin: 'com.android.library'
 apply plugin: 'com.jfrog.bintray'
 
 project.ext {
-    libVersion = '1.6'
-    libVersionCode = 16
+    libVersion = '0.1'
+    libVersionCode = 1
     libGroupId = 'com.xiaofeng.android'
     libArtifactId = 'flowlayoutmanager'
     gitUrl = 'git@github.com:xiaofeng-han/AndroidLibs.git'
@@ -21,13 +21,12 @@ version = libVersion
 
 android {
     compileSdkVersion androidCompileSdkVersion
-    resourcePrefix "xiaofeng__"
     defaultConfig {
         minSdkVersion androidMinSdkVersion
         targetSdkVersion androidTargetSdkVersion
         versionCode libVersionCode
         versionName libVersion
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     testOptions {
         unitTests.returnDefaultValues = false
@@ -39,20 +38,14 @@ android {
         }
     }
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
-    buildToolsVersion '27.0.3'
 }
 
 dependencies {
-    implementation fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestImplementation 'junit:junit:4.12'
-    implementation(group: 'com.android.support', name: 'appcompat-v7', version: supportLibVersion)
-    implementation(group: 'com.android.support', name: 'recyclerview-v7', version: supportLibVersion)
-    androidTestImplementation 'com.android.support:support-annotations:27.1.1'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-
+    implementation(group: 'androidx.appcompat', name: 'appcompat', version: supportLibVersion)
+    implementation(group: 'androidx.recyclerview', name: 'recyclerview', version: supportLibVersion)
 }
 
 Properties properties = new Properties()
