@@ -20,8 +20,8 @@ package com.xiaofeng.flowlayoutmanager;
  * Created by Anton Dudakov (dude.bw@gmail.com) on 13/10/16.
  */
 
-import android.support.v7.widget.OrientationHelper;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.OrientationHelper;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 /**
